# Remade TestSuite and TestGroup classes.