--- a/benchmarks/GEOMETRY_circle_benchmark.py
+++ b/benchmarks/GEOMETRY_circle_benchmark.py
@@ -2,8 +2,6 @@ from pygame import Rect
 from benchmark_utils import TestSuite
 from geometry import Circle
 
-CPT = 1_000_000  # Calls Per Test
-
 r1 = Rect(0, 0, 10, 10)
 r2 = Rect(10, 10, 4, 4)
 r3 = Rect(0, 0, 100, 100)
@@ -34,8 +32,27 @@ GLOB = {
 }
 
 # === Tests ===
-general_test = [
-    ("Instatiation", "Circle(0, 0, 5)"),
+# Each test consists of a tuple of: (name, call)
+# The name is a string containing the name of the test
+# The call is a string containing the code to be timed
+# every test is run CPT times and the average time is
+# calculated across REP runs
+# the formula is time = (time per CPT calls repeated REP times) / REP
+# ====================================================
+creation_tests = [
+    ("Circle int", "Circle(0, 0, 5)"),
+    ("Circle float", "Circle(0.0, 0.0, 5.0)"),
+]
+
+copy_tests = [
+    ("Circle copy", "c1.copy()"),
+]
+
+conversion_tests = [
+    ("as_rect", "c1.as_rect()"),
+]
+
+attributes_tests = [
     ("x get", "c1.x"),
     ("x set int", "c1.x = 3"),
     ("x set float", "c1.x = 3.0"),
@@ -54,33 +71,38 @@ general_test = [
     ("circumference get", "c1.circumference"),
     ("circumference set int", "c1.circumference = 3"),
     ("circumference set float", "c1.circumference = 3.0"),
-    ("copy", "c1.copy()"),
-    ("update int", "c1.update(1, 1, 3)"),
-    ("update float", "c1.update(1.0, 1.0, 3.0)"),
-    # ("move", "c1.move(1, 1)"),
-    # ("move_ip", "c1.move_ip(1, 1)"),
-    # ("move_to", "c1.move_to(1, 1)"),
-    # ("move_to_ip", "c1.move_to_ip(1, 1)"),
-    # ("scale_by", "c1.scale_by(1.2)"),
-    # ("scale_by_ip", "c1.scale_by_ip(1.2)"),
-    ("as_rect", "c1.as_rect()"),
 ]
 
-CC_collision_test = [
+update_tests = [
+    ("update circle", "c1.update(c2)"),
+    ("update 3 int", "c1.update(1, 1, 3)"),
+    ("update 3 float", "c1.update(1.0, 1.0, 3.0)"),
+    ("update 1 tup int", "c1.update((1, 1, 3))"),
+    ("update 1 tup float", "c1.update((1.0, 1.0, 3.0))"),
+]
+
+move_tests = [
+    ("move int", "c1.move(1, 1)"),
+    ("move float", "c1.move(1.0, 1.0)"),
+    ("move_ip int", "c1.move_ip(1, 1)"),
+    ("move_ip float", "c1.move_ip(1.0, 1.0)"),
+]
+
+CC_collision_tests = [
     ("Colliding", "c1.collidecircle(c2)"),
     ("Non colliding", "c1.collidecircle(c3)"),
     ("A inside B", "c1.collidecircle(c4)"),
     ("B inside A", "c4.collidecircle(c1)"),
 ]
 
-CR_collision_test = [
+CR_collision_tests = [
     ("Colliding", "c1.colliderect(r1)"),
     ("Non colliding", "c3.colliderect(r1)"),
     ("A inside B", "c3.colliderect(r3)"),
     ("B inside A", "c1.colliderect(r2)"),
 ]
 
-CP_collision_test = [
+CP_collision_tests = [
     ("Colliding 1 int", "c1.collidepoint(p1)"),
     ("Non colliding 1 int", "c1.collidepoint(p2)"),
     ("Colliding 1 float", "c1.collidepoint(p3)"),
@@ -91,7 +113,7 @@ CP_collision_test = [
     ("Non colliding 2 float", "c1.collidepoint(1000.0, 1000.0)"),
 ]
 
-CS_collision_test = [
+CS_collision_tests = [
     ("RECT colliding", "c1.collideswith(r1)"),
     ("RECT non colliding", "c1.collideswith(r2)"),
     ("CIRCLE colliding", "c1.collideswith(c2)"),
@@ -100,13 +122,20 @@ CS_collision_test = [
     ("POINT non colliding", "c1.collideswith(p2)"),
 ]
 
-TESTS = [
-    ("General", general_test),
-    ("Circle-Circle", CC_collision_test),
-    ("Circle-Rect", CR_collision_test),
-    ("Circle-Point", CP_collision_test),
-    ("Circle-Shape", CS_collision_test),
+# === Test Suites ===
+# If you want to add more tests to a suite, just add them to the list
+# If you want to remove or skip tests from a suite, just remove or comment them out
+GROUPS = [
+    ("Creation", creation_tests),
+    ("Attributes", attributes_tests),
+    ("Copy", copy_tests),
+    ("Conversion", conversion_tests),
+    ("Update", update_tests),
+    # ("Move", move_tests),
+    ("Circle-Circle", CC_collision_tests),
+    ("Circle-Rect", CR_collision_tests),
+    ("Circle-Point", CP_collision_tests),
+    ("Circle-Shape", CS_collision_tests),
 ]
-# ==========================
 
-TestSuite("Geometry Module - Circle", TESTS, GLOB, num=CPT).run_suite()
+TestSuite("Geometry Module - Circle", GROUPS, GLOB).run_suite()