--- a/requirements.txt
+++ b/requirements.txt
@@ -1,0 +1,3 @@
+pygame~=2.1.3.dev5
+setuptools~=63.2.0
+prettytable~=3.3.0