Now includes:
- Standard deviation for each test and TestGroup
- Median time for TestGroup
- Total time for each test and TestGroup
- Customizable time format (like ms, ns, us)
- Customizable precision
- Better formatting

Also:
- Added some more circle tests

**BENCHMARK OUTPUT BEFORE THIS PR:**
```
=================================
     GEOMETRY MODULE - LINE
=================================
Number of tests: 1000000 | Repeat: 5
=================================

====||GENERAL||====
-) Instatiation: 0.10190114 s
-) x1 attrib: 0.0279783 s
-) x1 attrib set: 0.0412057 s
-) y1 attrib: 0.0284062 s
-) y1 attrib set: 0.04023876 s
-) x2 attrib: 0.02841934 s
-) x2 attrib set: 0.0409867 s
-) y2 attrib: 0.02828576 s
-) y2 attrib set: 0.04046404 s
-) a attrib: 0.04668834 s
-) a attrib set: 0.06491622 s
-) b attrib: 0.04721828 s
-) b attrib set: 0.06550226 s
-) copy: 0.0439536 s
-) update line: 0.03712862 s
-) update 1 tup: 0.07677354 s
-) update 1 tup 2 subtups: 0.10370312 s
-) update 2 args: 0.11493824 s
-) update 4 args: 0.08243086 s
-) as_rect: 0.05785092 s
-----------------
Group Mean: 0.0559495

====||COLLISION: LINE-LINE||====
-) Colliding: 0.04295886 s
-) Non colliding: 0.04546354 s
-----------------
Group Mean: 0.0442112

====||COLLISION: LINE-CIRCLE||====
-) Colliding: 0.04184478 s
-) Non colliding: 0.05352914 s
-) inside circle: 0.04417386 s
-----------------
Group Mean: 0.04651593

====||COLLISION: LINE-POINT||====
-) Colliding: 0.0640266 s
-) Non colliding: 0.08539496 s
-) Colliding 2: 0.04161752 s
-) Non colliding 2: 0.06209112 s
-----------------
Group Mean: 0.06328255

====||RAYCAST||====
-) raycast: 1.09643702 s
-----------------
Group Mean: 1.09643702
=================================

Test suite total time: 13.482636700007792
```

**BENCHMARK OUTPUT AFTER THIS PR:**

```
=================================
     GEOMETRY MODULE - LINE
=================================
Test suite settings:
- Tests Number: 1000000
- Repeat Number: 5
- Time Format: 'ms'  (available 's', 'ms', 'us', 'ns')
- Precision: 2
- Single test total: Yes
- Single test mean: Yes
- Single test standard dev: Yes
- Current data order: TOTAL, MEAN, STD

==================================================
                     CREATION

|    Name    |  Total  |  Mean  |  Std Dev  |
|------------|---------|--------|-----------|
|  Line int  | 526.09  | 105.22 |   1.68    |
| Line float | 366.12  | 73.22  |   1.91    |

--------------------------------------------------
Group Mean: 89.22
Group Total: 892.21
Group Standard Deviation: 16.0
Group Median: 89.22



==================================================
                    ATTRIBUTES

|     Name     |  Total  |  Mean  |  Std Dev  |
|--------------|---------|--------|-----------|
|    x1 get    | 140.86  | 28.17  |   0.28    |
|  x1 set int  | 203.03  | 40.61  |   0.13    |
| x1 set float | 142.72  | 28.54  |   0.44    |
|    y1 get    | 142.98  |  28.6  |   0.51    |
|  y1 set int  | 203.81  | 40.76  |   0.54    |
| y1 set float | 141.37  | 28.27  |   0.29    |
|    x2 get    | 143.42  | 28.68  |    0.6    |
|  x2 set int  | 207.76  | 41.55  |   0.35    |
| x2 set float | 143.51  |  28.7  |   0.37    |
|    y2 get    | 142.59  | 28.52  |   0.42    |
|  y2 set int  | 204.12  | 40.82  |    1.1    |
| y2 set float | 140.51  |  28.1  |   0.39    |
|    a get     | 234.24  | 46.85  |   0.51    |
|  a set int   | 324.07  | 64.81  |   0.44    |
| a set float  | 234.75  | 46.95  |   0.79    |
|    b get     | 238.35  | 47.67  |   1.11    |
|  b set int   | 337.48  |  67.5  |   0.54    |
| b set float  | 233.89  | 46.78  |    0.8    |

--------------------------------------------------
Group Mean: 39.55
Group Total: 3559.46
Group Standard Deviation: 12.11
Group Median: 40.68



==================================================
                       COPY

|  Name  |  Total  |  Mean  |  Std Dev  |
|--------|---------|--------|-----------|
|  copy  |  224.5  |  44.9  |   0.93    |

--------------------------------------------------
Group Mean: 44.9
Group Total: 224.5
Group Standard Deviation: 0.0
Group Median: 44.9



==================================================
                    CONVERSION

|  Name   |  Total  |  Mean  |  Std Dev  |
|---------|---------|--------|-----------|
| as_rect | 291.45  | 58.29  |   1.39    |

--------------------------------------------------
Group Mean: 58.29
Group Total: 291.45
Group Standard Deviation: 0.0
Group Median: 58.29



==================================================
                      UPDATE

|          Name          |  Total  |  Mean  |  Std Dev  |
|------------------------|---------|--------|-----------|
|      update line       | 184.61  | 36.92  |   0.16    |
|      update 1 tup      | 382.03  | 76.41  |   0.47    |
| update 1 tup 2 subtups | 519.76  | 103.95 |   1.08    |
|     update 2 args      | 569.96  | 113.99 |    0.6    |
|     update 4 args      | 420.02  |   84   |   1.24    |

--------------------------------------------------
Group Mean: 83.06
Group Total: 2076.39
Group Standard Deviation: 26.71
Group Median: 84.0



==================================================
               COLLISION: LINE-LINE

|     Name      |  Total  |  Mean  |  Std Dev  |
|---------------|---------|--------|-----------|
|   Colliding   | 216.65  | 43.33  |    0.3    |
| Non colliding | 216.28  | 43.26  |   0.51    |

--------------------------------------------------
Group Mean: 43.29
Group Total: 432.92
Group Standard Deviation: 0.04
Group Median: 43.29



==================================================
              COLLISION: LINE-CIRCLE

|     Name      |  Total  |  Mean  |  Std Dev  |
|---------------|---------|--------|-----------|
|   Colliding   | 224.77  | 44.95  |   1.15    |
| Non colliding | 254.14  | 50.83  |   0.53    |
| inside circle | 205.35  | 41.07  |   0.45    |

--------------------------------------------------
Group Mean: 45.62
Group Total: 684.26
Group Standard Deviation: 4.01
Group Median: 44.95



==================================================
              COLLISION: LINE-POINT

|      Name       |  Total  |  Mean  |  Std Dev  |
|-----------------|---------|--------|-----------|
|    Colliding    | 306.36  | 61.27  |    0.4    |
|  Non colliding  | 402.15  | 80.43  |     1     |
|   Colliding 2   | 214.79  | 42.96  |   1.57    |
| Non colliding 2 | 312.85  | 62.57  |   1.29    |

--------------------------------------------------
Group Mean: 61.81
Group Total: 1236.16
Group Standard Deviation: 13.26
Group Median: 61.92



==================================================
                     RAYCAST

|  Name   |  Total  |  Mean  |  Std Dev  |
|---------|---------|--------|-----------|
| raycast | 1624.33 | 324.87 |   2.99    |

--------------------------------------------------
Group Mean: 324.87
Group Total: 1624.33
Group Standard Deviation: 0.0
Group Median: 324.87



_________________________________

Test suite total time: 11021.67 ms
```